Priority metadata support both int and enum values. Support for sync and async behavior rules. RequiredAnnotation.

+++ src/core/Kephas.Model/Runtime/Factory/Annotations/RuntimeRequiredAnnotationInfoFactory.cs
@@ -16,7 +16,7 @@ namespace Kephas.Model.Runtime.Factory.Annotations
     /// <summary>
     /// A runtime factory for the required annotation information.
     /// </summary>
-    [ProcessingPriority((int)Priority.AboveNormal)]
+    [ProcessingPriority(Priority.AboveNormal)]
     public class RuntimeRequiredAnnotationInfoFactory : RuntimeAnnotationInfoFactoryBase<RuntimeRequiredAnnotationInfo, RequiredAttribute>
     {
         /// <summary>
